Breaking the signed int glass ceiling; stage 1: convert critical ints to longs.  Code cleanup and documentation.
git-svn-id: file:///humgen/gsa-scr1/gsa-engineering/svn_contents/trunk@1852 348d0f76-0448-11de-a6fe-93d51630548a

+++ java/src/org/broadinstitute/sting/alignment/bwa/BWAAligner.java
@@ -124,7 +124,7 @@ public class BWAAligner implements Aligner {
 
             // Found a valid alignment; store it and move on.
             if(alignment.position >= read.getReadLength()-1) {
-                for( int bwtIndex = alignment.loBound; bwtIndex <= alignment.hiBound; bwtIndex++ ) {
+                for(long bwtIndex = alignment.loBound; bwtIndex <= alignment.hiBound; bwtIndex++) {
                     BWAAlignment finalAlignment = alignment.clone();
 
                     if( finalAlignment.isNegativeStrand() )
