+++ java/src/org/broadinstitute/sting/alignment/bwa/bwt/SuffixArray.java
@@ -9,13 +9,13 @@ import org.broadinstitute.sting.utils.StingException;
  * @version 0.1
  */
 public class SuffixArray {
-    public final int inverseSA0;
+    public final long inverseSA0;
     public final Counts occurrences;
 
     /**
      * The elements of the sequence actually stored in memory.
      */
-    protected final int[] sequence;
+    protected final long[] sequence;
 
     /**
      * How often are individual elements in the sequence actually stored
@@ -28,7 +28,7 @@ public class SuffixArray {
      */
     protected final BWT bwt;
 
-    public SuffixArray(int inverseSA0, Counts occurrences, int[] sequence) {
+    public SuffixArray(long inverseSA0, Counts occurrences, long[] sequence) {
         this(inverseSA0,occurrences,sequence,1,null);
     }
 
@@ -37,8 +37,10 @@ public class SuffixArray {
      * @param inverseSA0 Inverse SA entry for the first element.
      * @param occurrences Cumulative number of occurrences of A,C,G,T, in order.
      * @param sequence The full suffix array.
+     * @param sequenceInterval How frequently is the sequence interval stored.
+     * @param bwt bwt used to infer the remaining entries in the BWT.
      */
-    public SuffixArray(int inverseSA0, Counts occurrences, int[] sequence, int sequenceInterval, BWT bwt) {
+    public SuffixArray(long inverseSA0, Counts occurrences, long[] sequence, int sequenceInterval, BWT bwt) {
         this.inverseSA0 = inverseSA0;
         this.occurrences = occurrences;
         this.sequence = sequence;
@@ -53,7 +55,7 @@ public class SuffixArray {
      * Retrieves the length of the sequence array.
      * @return Length of the suffix array.
      */
-    public int length() {
+    public long length() {
         if( bwt != null )
             return bwt.length()+1;
         else
@@ -65,7 +67,7 @@ public class SuffixArray {
      * @param index Index at which to retrieve the suffix array vaule.
      * @return The suffix array value at that entry.
      */
-    public int get(int index) {
+    public long get(long index) {
         int iterations = 0;
         while(index%sequenceInterval != 0) {
             // The inverseSA0 ('$') doesn't have a usable ASCII representation; it must be treated as a special case.
@@ -77,6 +79,6 @@ public class SuffixArray {
             }
             iterations++;
         }
-        return (sequence[index/sequenceInterval]+iterations) % length();
+        return (sequence[(int)(index/sequenceInterval)]+iterations) % length();
     }
 }
